fix on tracking counter
